# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Sparkwing Controller HTTP API
  version: "0.1"
  description: |
    HTTP API served by `sparkwing-controller` (cluster mode) and embedded
    in the `sparkwing` CLI's `serve start` supervisor (laptop mode).
    Same route table, same wire shapes; the mode is determined by which
    functional options the process wires into the
    [pkg/controller.Server][1] instance.

    Mode-conditional routes are marked in each operation's `description`:

    - **cluster-only** -- wired when the process calls `AttachPool`
      (warm-PVC pool routes under `/api/v1/pool/*`).
    - **laptop-only** -- wired when the process calls `WithArtifactStore`
      (`GET /api/v1/artifacts/{key}`).

    See [VERSIONING.md][2] for stability promises around this surface:
    JSON field renames and behavioral changes are tracked under module
    semver. The Go client lives at `pkg/controller/client/`.

    [1]: https://pkg.go.dev/github.com/sparkwing-dev/sparkwing/pkg/controller
    [2]: ../VERSIONING.md
  contact:
    name: sparkwing
    url: https://github.com/sparkwing-dev/sparkwing
  license:
    name: Elastic License 2.0
    url: https://github.com/sparkwing-dev/sparkwing/blob/main/LICENSE
servers:
  - url: http://localhost:4344
    description: Default cluster-mode `sparkwing-controller` bind address.
  - url: http://127.0.0.1:4343
    description: Default laptop-mode `sparkwing serve start` bind address.
...
security:
  - bearerToken: []
paths:
...
  /api/v1/gitcache/git/register:
    post:
      x-sparkwing-scope: admin
      tags: [gitcache]
      summary: Map a short cache name onto a repository URL.
      parameters:
        - in: query
          name: name
          required: true
          schema: {type: string, pattern: "^[A-Za-z0-9._-]{1,64}$"}
        - {$ref: "#/components/parameters/GitcacheRepo"}
      responses:
        "200": {description: "The cache's answer, forwarded verbatim."}
        "400": {description: Rejected name or repo URL.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "502": {description: The cache is unreachable or answered badly.}
        "503": {description: No cache is configured on this controller.}
  /api/v1/gitcache/git/{path}:
    parameters:
      - in: path
        name: path
        required: true
        schema: {type: string}
        description: |
          `<registered-name>/info/refs` or `<registered-name>/git-upload-pack`.
          Registered by `POST /api/v1/gitcache/git/register`.
    get:
      x-sparkwing-scope: admin
      tags: [gitcache]
      summary: Advertise a mirror's refs to a Git client.
      description: |
        Forwards `info/refs?service=git-upload-pack` to the cache. Any other
        suffix or service answers `405`, so the proxy carries reads only.
        The controller drops the caller's bearer and presents its own cache
        token, and caps how many Git streams one caller may hold open.
      parameters:
        - in: query
          name: service
          required: true
          schema: {type: string, enum: [git-upload-pack]}
      responses:
        "200": {description: Git smart-protocol ref advertisement.}
        "400": {description: The path does not name a registered mirror.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "405": {description: Anything but an upload-pack ref advertisement.}
        "502": {description: The cache is unreachable or answered badly.}
        "503": {description: No cache is configured on this controller.}
    post:
      x-sparkwing-scope: admin
      tags: [gitcache]
      summary: Serve a Git fetch negotiation from a mirror.
      description: |
        Forwards `git-upload-pack` to the cache. Any other suffix answers
        `405`, so a caller cannot push through this proxy.
      requestBody:
        required: true
        content:
          application/x-git-upload-pack-request:
            schema: {type: string, format: binary}
      responses:
        "200": {description: Git smart-protocol packfile response.}
        "400": {description: The path does not name a registered mirror.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "405": {description: Anything but an upload-pack fetch.}
        "502": {description: The cache is unreachable or answered badly.}
        "503": {description: No cache is configured on this controller.}
...
  /api/v1/runs/{id}/gitcache/git/register:
    parameters:
      - in: path
        name: id
        required: true
        schema: {type: string}
    post:
      x-sparkwing-scope: nodes.claim
      tags: [gitcache]
      summary: Map the claimed run's source repository into the cache.
      description: |
        Run-scoped counterpart of `/api/v1/gitcache/git/register`. A
        claim-only caller must hold this run, and `name` plus `repo` must
        identify the run's recorded source. The controller forwards neither
        the caller's bearer nor arbitrary repository credentials to the cache.
      parameters:
        - in: query
          name: name
          required: true
          schema: {type: string, pattern: "^[A-Za-z0-9._-]{1,64}$"}
        - {$ref: "#/components/parameters/GitcacheRepo"}
      responses:
        "200": {description: "The cache's answer, forwarded verbatim."}
        "400": {description: Rejected name or repository URL.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {description: Caller does not hold the run or the repository is not its source.}
        "502": {description: The cache is unreachable or answered badly.}
        "503": {description: No cache is configured on this controller.}
  /api/v1/runs/{id}/gitcache/git/{path}:
    parameters:
      - in: path
        name: id
        required: true
        schema: {type: string}
      - in: path
        name: path
        required: true
        schema: {type: string}
    get:
      x-sparkwing-scope: nodes.claim
      tags: [gitcache]
      summary: Advertise the claimed run's cached source refs.
      description: |
        Forwards only `<registered-name>/info/refs?service=git-upload-pack`.
        A claim-only caller must hold this run, and the registered name must
        identify its recorded source repository.
      parameters:
        - in: query
          name: service
          required: true
          schema: {type: string, enum: [git-upload-pack]}
      responses:
        "200": {description: Git smart-protocol ref advertisement.}
        "400": {description: The path does not name a valid registered mirror.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {description: Caller does not hold the run or the mirror is not its source.}
        "405": {description: Anything but an upload-pack ref advertisement.}
        "502": {description: The cache is unreachable or answered badly.}
        "503": {description: No cache is configured on this controller.}
    post:
      x-sparkwing-scope: nodes.claim
      tags: [gitcache]
      summary: Fetch from the claimed run's cached source mirror.
      description: |
        Forwards only `<registered-name>/git-upload-pack`. A claim-only caller
        must hold this run, and the registered name must identify its recorded
        source repository. Push requests cannot cross this proxy.
      requestBody:
        required: true
        content:
          application/x-git-upload-pack-request:
            schema: {type: string, format: binary}
      responses:
        "200": {description: Git smart-protocol packfile response.}
        "400": {description: The path does not name a valid registered mirror.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {description: Caller does not hold the run or the mirror is not its source.}
        "405": {description: Anything but an upload-pack fetch.}
        "502": {description: The cache is unreachable or answered badly.}
        "503": {description: No cache is configured on this controller.}
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>` with a `sw*_`-prefixed token
        issued by `POST /api/v1/tokens`. The token's scopes (e.g.
        `runs.read`, `nodes.claim`, `admin`) gate per-route access.
    sessionAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        `Authorization: Session <session-id>`. Used only by the
        dashboard's browser flow after `POST /api/v1/auth/login`.
  parameters:
...
    GitcacheRepo:
      in: query
      name: repo
      required: true
      schema: {type: string}
      description: |
        Clone URL. Validated before proxying: a local path, a loopback or
        private address, and a URL carrying embedded credentials are all
        refused.
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/AuthError"}
...
  schemas:
...
    AuthError:
      x-sparkwing-go-type: pkg/controller.authErrorBody
      type: object
      description: Error returned by the auth middleware. Carries a missing-scope hint when applicable.
      required: [error]
      properties:
        error: {type: string}
        missing_scope: {type: string}
        principal: {type: string}
        message: {type: string}
...